Look at this further later - 64 bit fixes
git-svn-id: http://svn.inspircd.org/repository/branches/1_0_stable@4003 e03df62e-2008-0410-955e-edbf42e46eb7

+++ src/modules/m_cloaking.cpp
@@ -48,7 +48,7 @@ using namespace std;
 #define F4(x, y, z) (y ^ (x | ~z))
 #define MD5STEP(f,w,x,y,z,in,s) (w += f(x,y,z) + in, w = (w<<s | w>>(32-s)) + x)
 
-typedef unsigned long word32;
+typedef unsigned int word32; /* NOT unsigned long. We don't support 16 bit platforms, anyway. */
 typedef unsigned char byte;
 
 struct xMD5Context {
